move to `/internals` and disallow arrays
